@@ -1,13 +1,37 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
+import './index.css';
+import reportWebVitals from './reportWebVitals';
 import { ThemeProvider, createTheme } from '@mui/material/styles';
-import App from './App';
+import RouteProvider from './providers/RouterProvider.js';
+import { BrowserRouter } from 'react-router-dom';
+import Menu from './components/menu/Menu'
 
-const theme = createTheme();
+const theme = createTheme({
+    palette: {
+        primary: {
+            main: '#1976d2',
+        },
+        secondary: {
+            main: '#dc004e',
+        },
+    },
+});
 
-ReactDOM.render(
-  <ThemeProvider theme={theme}>
-      <App />
-  </ThemeProvider>,
-  document.getElementById('root')
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>
+      <ThemeProvider theme={theme}>
+          <BrowserRouter>
+              <Menu/>
+              <RouteProvider />
+          </BrowserRouter>
+      </ThemeProvider>
+  </React.StrictMode>
 );
+
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
